Refactor ModuleHeader into own component; rename 'moduleSlug' prop.
